### 2023-03-10 #5244 @editorialbot
[PRE REVIEW]: minorbsem: An R package for structural equation models that account for the influence of minor factors
**Submitting author:** @jamesuanhoro (<a href="http://orcid.org/0000-0002-4843-927X">James Ohisei Uanhoro</a>)
**Repository:** https://github.com/jamesuanhoro/minorbsem
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.1.0
