+++ inst/REFERENCES.bib
@@ -94,5 +94,6 @@
   author     = {{Inter-university Consortium for Political and Social Research}},
   publisher  = {Inter-university Consortium for Political and Social Research},
   year       = {1992},
+  doi        = {10.4232/1.1840},
   note       = {Google-Books-ID: bFK7AAAAIAAJ}
 }
