+++ R/meta_mbcfa.R
@@ -19,6 +19,8 @@
 #' "GDP", or "none". See details below.
 #' @inheritParams minorbsem
 #' @returns An object of \code{\link{mbsem-class}}
+#' Note: CFAs assume standardized factors.
+#' Will not fit latent variable regression models.
 #' @details
 #' There are different methods for estimating models in this package:
 #'
@@ -30,7 +32,8 @@
 #' accomodates extreme outliers.
 #' - \code{GDP}: to mimic a global-local approach, i.e.
 #' attempt to shrink near 0 residual covariances to 0
-#' with minimal shrinking for larger residual covariances.
+#' with minimal shrinking for larger residual covariances
+#' \insertCite{armagan_generalized_2013}{minorbsem}.
 #' - \code{none}: if intending to ignore the influence of minor factors.
 #' @examples
 #' \dontrun{
