+++ R/minorbsem.R
@@ -38,9 +38,10 @@
 #' @param show_messages (Logical) If TRUE, show messages from Stan sampler,
 #' if FALSE, hide messages.
 #' @returns An object of \code{\link{mbsem-class}}.
-#' Note: CFAs assume standardized factors.
-#' Latent variable regression models print results with standardized loadings.
 #' @details
+#' CFAs assume standardized factors.
+#' Latent variable regression models print results with standardized loadings.
+#'
 #' There are different methods for estimating models in this package:
 #'
 #' - \code{normal}: under belief that minor factor influences are
