+++ DESCRIPTION
@@ -26,7 +26,7 @@ Imports:
     Rcpp,
     Rdpack (>= 0.7)
 Remotes: 
-    github::stan-dev/cmdstanr
+    github::stan-dev/cmdstanr,
     github::stan-dev/posterior
 LazyData: true
 Suggests: 
