### 2023-03-17 #5244 @samhforbes
:wave: @smasongarrison, @aaronpeikert would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
