+++ README.Rmd
@@ -26,7 +26,7 @@ devtools::load_all()
   - [Goals](#goals)
   - [Permitted models and supported data types](#permitted-models-and-supported-data-types)
 - [Installation](#installation)
-  - [Install devtools](#install-devtools)
+  - [Install remotes](#install-remotes)
   - [Install CmdStanR and CmdStan](#install-cmdstanr-and-cmdstan)
   - [Install minorbsem](#install-minorbsem)
 - [A reasonably complete demonstration](#a-reasonably-complete-demonstration)
@@ -79,14 +79,14 @@ see installation instructions here: https://mc-stan.org/cmdstanr/articles/cmdsta
 
 If you have CmdStan installed and working, go to: [Install minorbsem](#install-minorbsem).
 
-### Install devtools
+### Install remotes
 
 Restart R prior to installation. minorbsem requires that R is at least version 3.4.0. `version` command shows the installed R version. If the installed version is less than 3.4.0, then update R before proceeding.
 
-Install `devtools` to allow installation of R packages that are hosted on GitHub instead of Cran (the default repository for R packages):
+Install `remotes` to allow installation of R packages that are hosted on GitHub instead of Cran (the default repository for R packages):
 
 ``` r
-install.packages("devtools")
+install.packages("remotes")
 ```
 
 ### Install CmdStanR and CmdStan
@@ -94,7 +94,7 @@ install.packages("devtools")
 Install the `cmdstanr` package so `minorbsem` can interface with Stan. If any problems come up at this stage, see: [Getting started with CmdStanR](https://mc-stan.org/cmdstanr/articles/cmdstanr.html).
 
 ``` r
-devtools::install_github("stan-dev/cmdstanr")
+remotes::install_github("stan-dev/cmdstanr")
 ```
 
 Install Stan. First, we check that the computer is correctly set up to install Stan:
@@ -109,7 +109,7 @@ cmdstanr::install_cmdstan()  # install Stan
 Install minorbsem:
 
 ``` r
-devtools::install_github("jamesuanhoro/minorbsem")
+remotes::install_github("jamesuanhoro/minorbsem")
 ```
 
 ## A reasonably complete demonstration
