+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 2.19.2
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles: {}
-last_built: 2023-03-11T15:32Z
+last_built: 2023-03-17T17:45Z
 urls:
   reference: https://jamesuanhoro.github.io/minorbsem/reference
   article: https://jamesuanhoro.github.io/minorbsem/articles
