### 2023-03-20 6f69ff3
Added on load message
- Also using read.dcf to read DESCRIPTION file
