+++ man/type_hash.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils.R
+\name{type_hash}
+\alias{type_hash}
+\title{Type-meta hash function}
+\usage{
+type_hash(search_term = NULL, elaborate = FALSE)
+}
+\arguments{
+\item{search_term}{Integer or string or NULL}
+
+\item{elaborate}{(LOGICAL) If TRUE, print full names, otherwise
+print abbreviations}
+}
+\value{
+If search_term is integer, returns string and vice-versa
+}
+\description{
+A function that swaps type from string to integer
+and vice-versa
+}
+\keyword{internal}
