+++ NEWS.md
@@ -1,4 +1,10 @@
+# minorbsem 0.1.1
+
+* Added Wishart-based fixed-effects meta-analytic SEM (also permits minor factors)
+
 # minorbsem 0.1.0
 
 * First presentation of package
+  * CFA/SEM with minor factors assumed
+  * Random-effects meta-analytic permitting minor factor influences
 * Added a `NEWS.md` file to track changes to the package.
