+++ R/create_data_list_meta.R
@@ -1,7 +1,7 @@
 #' Stan data helper function
 #' @description A function that creates data list object passed to Stan
 #' @param lavaan_object lavaan fit object of corresponding model
-#' @inheritParams minorbsem
+#' @inheritParams meta_mbcfa
 #' @returns Data list object used in fitting Stan model
 #' @keywords internal
 create_data_list_meta <- function(
