### 2023-03-21 dc974fe
Also testing covariance inputs
