+++ tests/testthat/setup.R
@@ -57,3 +57,18 @@ mbsem_test_kbls_shared <- function(kbl, method, meta = FALSE) {
     testthat::expect_true(regexpr("PPP", kbl, ignore.case = TRUE) > 0)
   }
 }
+
+dat_cov <- function(dat = "HS") {
+  if (dat == "HS") dat <- HS[, paste0("x", 1:9)]
+  else if (dat == "PD") dat <- PD
+
+  input <- list(data = NULL, cov = NULL, nobs = NULL)
+  if (sample(2, 1) == 1) {
+    input$data <- dat
+  } else {
+    input$cov <- cov(dat)
+    input$nobs <- nrow(dat)
+  }
+
+  return(input)
+}
