### 2023-03-28 a4d2d38
More secure approach with 1 chain
