+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: minorbsem
 Title: Fit Bayesian structural equation models under the assumption that minor factors exist
-Version: 0.1.1.9000
+Version: 0.1.1.9001
 Authors@R: c(
     person(given = c("James", "O."), family = "Uanhoro",
            email = "james.uanhoro@unt.edu", role = c("aut", "cre"),
