+++ R/minorbsem-package.R
@@ -1,0 +1,16 @@
+#' The 'minorbsem' package.
+#'
+#' @description A DESCRIPTION OF THE PACKAGE
+#'
+#' @docType package
+#' @name minorbsem-package
+#' @aliases minorbsem
+#' @useDynLib minorbsem, .registration = TRUE
+#' @import methods
+#' @import Rcpp
+#' @importFrom rstan sampling
+#'
+#' @references
+#' Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.21. https://mc-stan.org
+#'
+NULL
