+++ R/prepare_stan_plot_data.R
@@ -1,0 +1,164 @@
+#' Pre-plotting Stan fit helper function
+#' @description A function that cleans up model results to prepare for plotting
+#' @param object (mbsem) An object of \code{\link{mbsem-class}}
+#' returned by \code{\link{minorbsem}}.
+#' @returns A data.frame that is ready for plotting
+#' @keywords internal
+prepare_stan_plot_data <- function(object) {
+  result <- list()
+
+  data_list <- object@data_list
+  stan_fit <- object@stan_fit
+
+  indicator_labels <- rownames(data_list$loading_pattern)
+  factor_labels <- colnames(data_list$loading_pattern)
+
+  rms_result <- posterior::as_draws_df(
+    posterior::as_draws(stan_fit, variable = "rms_src")
+  )
+  colnames(rms_result)[1] <- "rm: rms_src"
+
+  rmsea_result <- matrix(nrow = nrow(rms_result), ncol = 0)
+  if (data_list$meta == 1) {
+    rmsea_result <- posterior::as_draws_df(
+      posterior::as_draws(stan_fit, variable = "rmsea_mn")
+    )
+    colnames(rmsea_result)[1] <- "rm: rmsea"
+  }
+
+  load_idxs <- paste0("Load_mat[", apply(which(
+    data_list$loading_pattern >= ifelse(data_list$complex_struc == 1, -999, 1),
+    arr.ind = TRUE
+  ), 1, paste0, collapse = ","), "]")
+  load_result <- posterior::as_draws_df(
+    posterior::as_draws(stan_fit, variable = load_idxs)
+  )
+  load_result <- rename_post_df_columns(
+    load_result, factor_labels, indicator_labels, "lo:", "=~",
+    TRUE, "Load_mat\\[\\d+,|\\]", "Load_mat\\[|,\\d+\\]"
+  )
+
+  rv_result <- posterior::as_draws_df(
+    posterior::as_draws(stan_fit, variable = "res_var")
+  )
+  rv_result <- rename_post_df_columns(
+    rv_result, indicator_labels, indicator_labels, "ev:", "~~",
+    TRUE, "res_var\\[|\\]", "res_var\\[|\\]"
+  )
+
+  rc_result <- matrix(nrow = nrow(rv_result), ncol = 0)
+  if (data_list$Nce > 0) {
+    rc_result <- posterior::as_draws_df(
+      posterior::as_draws(stan_fit, variable = "res_cor")
+    )
+    rc_result <- rename_post_df_columns(
+      rc_result, indicator_labels, indicator_labels, "rc:", "~~",
+      FALSE, data_list$error_mat[, 1], data_list$error_mat[, 2]
+    )
+  }
+
+  coef_result <- matrix(nrow = nrow(rv_result), ncol = 0)
+  phi_result <- matrix(nrow = nrow(rv_result), ncol = 0)
+  rsq_result <- matrix(nrow = nrow(rv_result), ncol = 0)
+  if (data_list$sem_indicator == 0) {
+    phi_result <- posterior::as_draws_df(
+      posterior::as_draws(stan_fit, variable = "phi_mat")
+    )
+    phi_duplicates <- duplicated(as.list(phi_result))
+    phi_result <- phi_result[!phi_duplicates]
+    phi_result <- rename_post_df_columns(
+      phi_result, factor_labels, factor_labels, "fc:", "~~",
+      TRUE, "phi_mat\\[\\d+,|\\]", "phi_mat\\[|,\\d+\\]"
+    )
+  } else if (data_list$sem_indicator == 1) {
+    # Get interfactor correlations
+    if (data_list$Nf_corr > 0) {
+      phi_result <- posterior::as_draws_df(
+        posterior::as_draws(stan_fit, variable = "phi_cor")
[…]
