+++ R/minorbsem.R
@@ -1,0 +1,204 @@
+#' Fit Bayesian SEMs with minor factors assumed
+#'
+#' @description The major function to fit models assuming the influence of
+#' minor factors \insertCite{uanhoro_modeling_2023}{minorbsem}.
+#' @param model A description of the user-specified model, lavaan syntax.
+#' @param data An optional data frame containing the observed variables used in
+#' the model.
+#' @param sample_cov (matrix) sample variance-covariance matrix.
+#' The rownames and/or colnames must contain the observed variable names.
+#' @param sample_nobs (positive integer) Number of observations if the full
+#' data frame
+#' is missing and only sample covariance matrix is given.
+#' @param method (character) One of "normal", "lasso", "logistic",
+#' "GDP", "WB", "WB-cond", or "none". See details below.
+#' @param orthogonal (logical) constrain factors orthogonal, must be TRUE to fit
+#' bifactor models.
+#' @param simple_struc (LOGICAL) Only relevant for CFAs.
+#' If TRUE: assume simple structure;
+#' If FALSE: estimate all cross-loadings using generalized
+#' double Pareto priors.
+#' @param seed (positive integer) seed, set to obtain replicable results.
+#' @param warmup (positive integer) The number of warmup iterations to run per
+#' chain.
+#' @param sampling (positive integer) The number of post-warmup iterations to
+#' run per chain, retained for inference.
+#' @param refresh (positive integer) How often to print the status
+#' of the sampler.
+#' @param adapt_delta (real in (0, 1)) Increase to resolve divergent
+#' transitions.
+#' @param max_treedepth (positive integer) Increase to resolve problems with
+#' maximum tree depth.
+#' @param chains (positive integer) The number of Markov chains to run.
+#' @param ncores (positive integer) The number of chains to run in parallel.
+#' @param priors An object of \code{\link{mbsempriors-class}}.
+#' See \code{\link{new_mbsempriors}} for more information.
+#' @param show (Logical) If TRUE, show table of results, if FALSE, do not
+#' show table of results. As an example, use FALSE for simulation studies.
+#' @param show_messages (Logical) If TRUE, show messages from Stan sampler,
+#' if FALSE, hide messages.
+#' @returns An object of \code{\link{mbsem-class}}
+#' @details
+#' CFAs assume standardized factors.
+#' Latent variable regression models print results with standardized loadings.
+#'
+#' There are different methods for estimating models in this package:
+#'
+#' - \code{normal}: under belief that minor factor influences are
+#' on average zero with continuous deviations away from zero
+#' \insertCite{uanhoro_modeling_2023}{minorbsem}.
+#' - \code{lasso}: under belief that minor factor influences are largely
+#' zero with a small number of non-zero residual covariances.
+#' - \code{logistic}: for similar belief as normal but more readily
+#' accomodates extreme outliers.
+#' - \code{GDP}: to mimic a global-local approach, i.e.
+#' attempt to shrink near 0 residual covariances to 0
+#' with minimal shrinking for larger residual covariances
+#' \insertCite{armagan_generalized_2013}{minorbsem}.
+#' - \code{WB}: to model the covariance matrix hierarchically under
+#' assumptions of adventitiuous error
+#' \insertCite{wu_quantifying_2015}{minorbsem};
[…]
