+++ man/meta_mbcfa.Rd
@@ -1,0 +1,143 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/meta_mbcfa.R
+\name{meta_mbcfa}
+\alias{meta_mbcfa}
+\title{Fit random-effects Bayesian meta-analytic CFAs with minor factors assumed.}
+\usage{
+meta_mbcfa(
+  model = NULL,
+  data = NULL,
+  group = NULL,
+  sample_cov = NULL,
+  sample_nobs = NULL,
+  method = "normal",
+  type = "re",
+  orthogonal = FALSE,
+  simple_struc = TRUE,
+  seed = 12345,
+  warmup = 1000,
+  sampling = 1000,
+  refresh = (warmup + sampling)/10,
+  adapt_delta = 0.9,
+  max_treedepth = 10,
+  chains = 4,
+  ncores = max(parallel::detectCores() - 2, 1),
+  priors = new_mbsempriors(),
+  show = TRUE,
+  show_messages = TRUE
+)
+}
+\arguments{
+\item{model}{A description of the user-specified model, lavaan syntax.}
+
+\item{data}{An optional data frame containing the observed variables used in
+the model.}
+
+\item{group}{An optional string identifying the grouping variable in
+the data object.}
+
+\item{sample_cov}{(list of matrices) sample variance-covariance matrices.
+The rownames and/or colnames must contain the observed variable names.
+For now, assumes there are no missing elements in the covariance matrices.}
+
+\item{sample_nobs}{(vector of positive integer) Number of observations
+for each study.}
+
+\item{method}{(character) One of "normal", "lasso", "logistic",
+"GDP", or "none". See details below.}
+
+\item{type}{(character) One of "fe" or "re" for fixed-effects MASEM
+or random-effects MASEM respectively.}
+
+\item{orthogonal}{(logical) constrain factors orthogonal, must be TRUE to fit
+bifactor models.}
+
+\item{simple_struc}{(LOGICAL) Only relevant for CFAs.
+If TRUE: assume simple structure;
+If FALSE: estimate all cross-loadings using generalized
+double Pareto priors.}
+
+\item{seed}{(positive integer) seed, set to obtain replicable results.}
+
+\item{warmup}{(positive integer) The number of warmup iterations to run per
+chain.}
+
+\item{sampling}{(positive integer) The number of post-warmup iterations to
+run per chain, retained for inference.}
+
+\item{refresh}{(positive integer) How often to print the status
+of the sampler.}
+
+\item{adapt_delta}{(real in (0, 1)) Increase to resolve divergent
+transitions.}
+
+\item{max_treedepth}{(positive integer) Increase to resolve problems with
+maximum tree depth.}
+
+\item{chains}{(positive integer) The number of Markov chains to run.}
+
+\item{ncores}{(positive integer) The number of chains to run in parallel.}
+
+\item{priors}{An object of \code{\link{mbsempriors-class}}.
+See \code{\link{new_mbsempriors}} for more information.}
+
+\item{show}{(Logical) If TRUE, show table of results, if FALSE, do not
+show table of results. As an example, use FALSE for simulation studies.}
+
+\item{show_messages}{(Logical) If TRUE, show messages from Stan sampler,
+if FALSE, hide messages.}
+}
+\value{
+An object of \code{\link{mbsem-class}}
+}
+\description{
+A function to fit fixed- or
+random-effects Bayesian meta-analytic CFAs
+with minor factors assumed \insertCite{uanhoro_hierarchical_2022}{minorbsem}.
+Builds off work by \insertCite{wu_quantifying_2015;textual}{minorbsem}.
[…]
