+++ man/new_mbsempriors.Rd
@@ -1,0 +1,65 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/00_class_defintions.R
+\name{new_mbsempriors}
+\alias{new_mbsempriors}
+\title{Set priors in package}
+\usage{
+new_mbsempriors(
+  lkj_shape = 2,
+  sl_par = 1,
+  rs_par = 2.5,
+  rc_par = 2,
+  sc_par = 1,
+  fc_par = 2,
+  mln_par = 2.5,
+  mlb_par = 1
+)
+}
+\arguments{
+\item{lkj_shape}{(positive real) The shape parameter of the LKJ-prior on the
+interfactor correlation matrix in confirmatory factor models.}
+
+\item{sl_par}{(positive real) The scale parameter of the
+Student-t(df = 3, loc = 0) prior on the hyper-parameter of the standard
+deviation of loadings.}
+
+\item{rs_par}{(positive real) The scale parameter of the
+Student-t(df = 3, loc = 0) prior on the residual standard deviations.}
+
+\item{rc_par}{(positive real) The shape parameter of the Beta(rc_par, rc_par)
+prior on the residual error correlations.}
+
+\item{sc_par}{(positive real) The scale parameter of the
+Student-t(df = 3, loc = 0) prior on the hyper-parameter of the standard
+deviations of coefficients; SD(coefs) vary by outcome.}
+
+\item{fc_par}{(positive real) The shape parameter of the Beta(rc_par, rc_par)
+prior on the inter-factor correlations in latent regression models.}
+
+\item{mln_par}{(positive real) The scale parameter of the
+Student-t(df = 3, loc = 0) prior on the
+log-scale meta-regression intercept.
+Relevant only for meta-analysis.}
+
+\item{mlb_par}{(positive real) The scale parameter of the
+Student-t(df = 3, loc = 0) prior on the
+log-scale meta-regression coefficients.
+Relevant only for meta-analysis.}
+}
+\value{
+An object of \code{\link{mbsempriors-class}}
+}
+\description{
+Modify default priors in package.
+}
+\examples{
+# Change LKJ shape parameter only
+custom_priors <- new_mbsempriors(lkj_shape = 1.0)
+\dontrun{
+minorbsem("F1 =~ x1 + x2 + x3
+           F2 =~ x4 + x5 + x6
+           F3 =~ x7 + x8 + x9", HS,
+  priors = custom_priors
+)
+}
+}
