+++ man/prepare_stan_plot_data.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/prepare_stan_plot_data.R
+\name{prepare_stan_plot_data}
+\alias{prepare_stan_plot_data}
+\title{Pre-plotting Stan fit helper function}
+\usage{
+prepare_stan_plot_data(object)
+}
+\arguments{
+\item{object}{(mbsem) An object of \code{\link{mbsem-class}}
+returned by \code{\link{minorbsem}}.}
+}
+\value{
+A data.frame that is ready for plotting
+}
+\description{
+A function that cleans up model results to prepare for plotting
+}
+\keyword{internal}
