+++ inst/REFERENCES.bib
@@ -1,0 +1,117 @@
+
+@article{uanhoro_modeling_2023,
+  title    = {Modeling misspecification as a parameter in {Bayesian} structural equation modeling},
+  abstract = {Accounting for model misspecification in Bayesian structural equation models is an active area of research. We present a uniquely Bayesian approach to misspecification that models the degree of misspecification as a parameter – a parameter akin to the correlation root mean squared residual. The misspecification parameter can be interpreted on its own terms as a measure of absolute model fit, and allows for comparing different models fit to the same data. By estimating the degree of misspecification simultaneously with structural parameters, the uncertainty about structural parameters reflect the degree of model misspecification. This results in a model that produces more reliable inference than extant Bayesian SEMs. Additionally, the approach estimates the residual covariance matrix which can be the basis for diagnosing misspecifications and updating a hypothesized model. These features are confirmed using simulation studies. Demonstrations with a variety of real world examples show additional properties of the approach.},
+  language = {en},
+  journal  = {Educational and Psychological Measurement},
+  author   = {Uanhoro, James Ohisei},
+  year     = {forthcoming}
+}
+
+@article{uanhoro_hierarchical_2022,
+  title    = {Hierarchical Covariance Estimation Approach to Meta-Analytic Structural Equation Modeling},
+  volume   = 0,
+  issn     = {1070-5511},
+  doi      = {10.1080/10705511.2022.2142128},
+  abstract = {We present a fully Bayesian approach to meta-analytic SEM based on hierarchical modeling of sample covariance matrices. The approach allows for flexible models that would not be identified under a traditional maximum likelihood approach. The approach allows for the inclusion of moderators, produces a global fit index, and permits the investigation of local misspecification. Simulation-based calibration studies show that the Bayesian computation procedure produces valid inferences for commonplace meta-analytic SEM applications. We demonstrate the approach with diverse data analysis examples and provide accompanying R code to support adoption and additional study of the approach. Finally, we lay out proposals that have the potential to extend the approach to accommodate the wide variety of analyses and data conditions that comprise meta-analytic SEM applications.},
+  number   = {0},
+  journal  = {Structural Equation Modeling: A Multidisciplinary Journal},
+  author   = {Uanhoro, James Ohisei},
+  month    = dec,
+  year     = {2022},
+  keywords = {Bayesian estimation, hierarchical covariance, meta-analytic SEM},
+  pages    = {1--15}
+}
+
+@article{maccallum_representing_1991,
+  title      = {Representing sources of error in the common-factor model: {Implications} for theory and practice},
+  volume     = {109},
+  issn       = {1939-1455},
+  shorttitle = {Representing sources of error in the common-factor model},
+  doi        = {10.1037/0033-2909.109.3.502},
[…]
