+++ DESCRIPTION
@@ -25,6 +25,7 @@ Imports:
     posterior,
     Rcpp (>= 0.12.0),
     RcppParallel (>= 5.0.1),
+    Rdpack,
     rstan (>= 2.18.1),
     rstantools (>= 2.3.0)
 LinkingTo: 
