+++ src/RcppExports.cpp
@@ -13,9 +13,13 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 
 
 RcppExport SEXP _rcpp_module_boot_stan_fit4cfa_resid_rs_mod();
+RcppExport SEXP _rcpp_module_boot_stan_fit4meta_cfa_resid_rs_mod();
+RcppExport SEXP _rcpp_module_boot_stan_fit4sem_resid_rs_mod();
 
 static const R_CallMethodDef CallEntries[] = {
     {"_rcpp_module_boot_stan_fit4cfa_resid_rs_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4cfa_resid_rs_mod, 0},
+    {"_rcpp_module_boot_stan_fit4meta_cfa_resid_rs_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4meta_cfa_resid_rs_mod, 0},
+    {"_rcpp_module_boot_stan_fit4sem_resid_rs_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4sem_resid_rs_mod, 0},
     {NULL, NULL, 0}
 };
 
