+++ R/stanmodels.R
@@ -1,10 +1,12 @@
 # Generated by rstantools.  Do not edit by hand.
 
 # names of stan models
-stanmodels <- c("cfa_resid_rs")
+stanmodels <- c("cfa_resid_rs", "meta_cfa_resid_rs", "sem_resid_rs")
 
 # load each stan module
 Rcpp::loadModule("stan_fit4cfa_resid_rs_mod", what = TRUE)
+Rcpp::loadModule("stan_fit4meta_cfa_resid_rs_mod", what = TRUE)
+Rcpp::loadModule("stan_fit4sem_resid_rs_mod", what = TRUE)
 
 # instantiate each stanmodel object
 stanmodels <- sapply(stanmodels, function(model_name) {
