+++ man/minorbsem-package.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/minorbsem-package.R
-\docType{package}
-\name{minorbsem-package}
-\alias{minorbsem-package}
-\alias{minorbsem}
-\title{The 'minorbsem' package.}
-\description{
-A DESCRIPTION OF THE PACKAGE
-}
-\references{
-Stan Development Team (NA). RStan: the R interface to Stan.
-R package version 2.26.21. https://mc-stan.org
-}
