+++ NAMESPACE
@@ -12,6 +12,7 @@ export(pretty_print_summary)
 exportClasses(mbsempriors)
 import(Rcpp)
 import(methods)
+importFrom(Rcpp,sourceCpp)
 importFrom(Rdpack,reprompt)
 importFrom(methods,show)
 importFrom(rstan,sampling)
