+++ DESCRIPTION
@@ -41,3 +41,4 @@ Suggests:
 Config/testthat/edition: 3
 RdMacros: Rdpack
 URL: https://jamesuanhoro.github.io/minorbsem/
+LazyData: true

[…]
