+++ README.md
@@ -1,0 +1,253 @@
+
+# minorbsem
+
+<!-- badges: start -->
+
+[![Project Status: Active The project has reached a stable, usable state
+and is being actively
+developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+![GitHub](https://img.shields.io/github/license/jamesuanhoro/minorbsem)
+[![Codecov test
+coverage](https://codecov.io/gh/jamesuanhoro/minorbsem/branch/master/graph/badge.svg)](https://app.codecov.io/gh/jamesuanhoro/minorbsem?branch=master)
+[![minimal R
+version](https://img.shields.io/badge/R%3E%3D-3.4.0-6666ff.svg)](https://cran.r-project.org/)
+![GitHub R package
+version](https://img.shields.io/github/r-package/v/jamesuanhoro/minorbsem)
+[![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/minorbsem)](https://cran.r-project.org/package=minorbsem)
+![GitHub last
+commit](https://img.shields.io/github/last-commit/jamesuanhoro/minorbsem)
+<!-- badges: end -->
+
+#### Table of Contents
+
+- [Package overview](#package-overview)
+  - [Goals](#goals)
+  - [Permitted models and supported data
+    types](#permitted-models-and-supported-data-types)
+- [Installation](#installation)
+  - [Install remotes](#install-remotes)
+  - [Install minorbsem](#install-minorbsem)
+- [A reasonably complete
+  demonstration](#a-reasonably-complete-demonstration)
+  - [Model comparisons](#model-comparisons)
+- [Additional examples](#additional-examples)
+  - [Different methods to capture the influence of minor
+    factors](#different-methods-to-capture-the-influence-of-minor-factors)
+  - [Meta-analytic CFA](#meta-analytic-cfa)
+- [Citations](#citations)
+
+## Package overview
+
+It is rare that structural equation models (SEMs) are able to reject the
+null hypothesis that there is no misspecification in the model. One
+explanation for this problem is that covariance structures are
+influenced by major factors which we can hypothesize about and minor
+factors which we cannot predict a-priori, e.g. MacCallum and Tucker
+(1991).
+
+### Goals
+
+The goal of minorbsem is to let you fit Bayesian SEMs that estimate the
+influence of minor factors on the covariance matrix following the
+approach in Uanhoro (forthcoming). Briefly, all residual covariances are
+estimated with priors that shrink them towards zero, and the model
+returns the magnitude of the influence of minor factors.
+
+minorbsem also fits random-effects meta-analytic CFAs that capture the
+influence of minor factors following the approach in Uanhoro (2022).
+
+### Permitted models and supported data types
+
+The package only fits a limited number of model configurations:
+
+- CFA, allowing cross-loadings (which may be automatically estimated),
+  correlated errors terms, and fully oblique or orthogonal factors
+  (useful for fitting bifactor models)
+- SEMs allowing latent regressions (only), cross-loadings, and
+  correlated error terms.
+
+One cannot fit MIMIC, multi-group models, multilevel models, or models
+with specially constrained parameters (e.g. setting two parameters
+equal).
+
+The meta-analysis models are only for the CFA configurations.
+
+All data are assumed multivariate normal, i.e. no binary, ordinal
+models.
+
+## Installation
+
[…]
