+++ DESCRIPTION
@@ -26,14 +26,14 @@ Imports:
     Rcpp (>= 0.12.0),
     RcppParallel (>= 5.0.1),
     Rdpack,
-    rstan (>= 2.18.1),
+    rstan (>= 2.26.21),
     rstantools (>= 2.3.0)
 LinkingTo: 
     BH (>= 1.66.0),
     Rcpp (>= 0.12.0),
     RcppEigen (>= 0.3.3.3.0),
     RcppParallel (>= 5.0.1),
-    rstan (>= 2.18.1),
+    rstan (>= 2.26.21),
     StanHeaders (>= 2.18.0)
 SystemRequirements: GNU make
 Suggests: 
@@ -42,3 +42,5 @@ Config/testthat/edition: 3
 RdMacros: Rdpack
 URL: https://jamesuanhoro.github.io/minorbsem/
 LazyData: true
+Additional_repositories:
+    https://mc-stan.org/r-packages/
