+++ src/stanExports_sem_resid_rs.h
@@ -75,413 +75,409 @@ using stan::math::pow;
 stan::math::profile_map profiles__;
 static int current_statement__= 0;
 static const std::vector<string> locations_array__ = {" (found before start of program)",
-                                                      " (in 'sem_resid_rs', line 89, column 2 to column 39)",
-                                                      " (in 'sem_resid_rs', line 90, column 2 to column 41)",
-                                                      " (in 'sem_resid_rs', line 91, column 2 to column 24)",
-                                                      " (in 'sem_resid_rs', line 92, column 2 to column 38)",
-                                                      " (in 'sem_resid_rs', line 93, column 2 to column 33)",
-                                                      " (in 'sem_resid_rs', line 94, column 2 to column 31)",
-                                                      " (in 'sem_resid_rs', line 95, column 2 to column 32)",
-                                                      " (in 'sem_resid_rs', line 96, column 2 to column 51)",
+                                                      " (in 'sem_resid_rs', line 87, column 2 to column 39)",
+                                                      " (in 'sem_resid_rs', line 88, column 2 to column 41)",
+                                                      " (in 'sem_resid_rs', line 89, column 2 to column 24)",
+                                                      " (in 'sem_resid_rs', line 90, column 2 to column 38)",
+                                                      " (in 'sem_resid_rs', line 91, column 2 to column 33)",
+                                                      " (in 'sem_resid_rs', line 92, column 2 to column 31)",
+                                                      " (in 'sem_resid_rs', line 93, column 2 to column 32)",
+                                                      " (in 'sem_resid_rs', line 94, column 2 to column 51)",
+                                                      " (in 'sem_resid_rs', line 95, column 2 to column 47)",
+                                                      " (in 'sem_resid_rs', line 96, column 2 to column 27)",
                                                       " (in 'sem_resid_rs', line 97, column 2 to column 47)",
-                                                      " (in 'sem_resid_rs', line 98, column 2 to column 27)",
-                                                      " (in 'sem_resid_rs', line 99, column 2 to column 47)",
-                                                      " (in 'sem_resid_rs', line 100, column 2 to column 28)",
-                                                      " (in 'sem_resid_rs', line 218, column 2 to column 13)",
-                                                      " (in 'sem_resid_rs', line 219, column 2 to column 13)",
-                                                      " (in 'sem_resid_rs', line 220, column 2 to column 33)",
-                                                      " (in 'sem_resid_rs', line 221, column 2 to column 26)",
-                                                      " (in 'sem_resid_rs', line 222, column 2 to column 52)",
-                                                      " (in 'sem_resid_rs', line 223, column 2 to column 52)",
-                                                      " (in 'sem_resid_rs', line 224, column 2 to column 50)",
-                                                      " (in 'sem_resid_rs', line 225, column 2 to column 50)",
-                                                      " (in 'sem_resid_rs', line 226, column 2 to column 38)",
-                                                      " (in 'sem_resid_rs', line 227, column 2 to column 22)",
-                                                      " (in 'sem_resid_rs', line 228, column 2 to column 47)",
-                                                      " (in 'sem_resid_rs', line 229, column 2 to column 39)",
-                                                      " (in 'sem_resid_rs', line 230, column 2 to column 43)",
-                                                      " (in 'sem_resid_rs', line 231, column 2 to column 22)",
-                                                      " (in 'sem_resid_rs', line 232, column 2 to column 49)",
-                                                      " (in 'sem_resid_rs', line 236, column 4 to column 18)",
-                                                      " (in 'sem_resid_rs', line 235, column 9 to line 237, column 3)",
-                                                      " (in 'sem_resid_rs', line 234, column 4 to column 27)",
[…]

[…]
