+++ DESCRIPTION
@@ -28,8 +28,7 @@ Imports:
     Rdpack,
     rstan (>= 2.21.8),
     rstantools (>= 2.3.0)
-Remotes: 
-    github::stan-dev/rstan@v2.26.21
+Remotes: url::https://github.com/stan-dev/rstan/archive/refs/tags/v2.26.21.zip
 LinkingTo: 
     BH (>= 1.66.0),
     Rcpp (>= 0.12.0),
