+++ src/stanExports_sem_resid_rs.h
@@ -162,7 +162,7 @@ static const std::vector<string> locations_array__ = {" (found before start of p
                                                       " (in 'sem_resid_rs', line 268, column 4 to column 25)",
                                                       " (in 'sem_resid_rs', line 269, column 11 to column 13)",
                                                       " (in 'sem_resid_rs', line 269, column 4 to column 26)",
-                                                      " (in 'sem_resid_rs', line 271, column 6 to column 43)",
+                                                      " (in 'sem_resid_rs', line 271, column 6 to column 37)",
                                                       " (in 'sem_resid_rs', line 273, column 40 to column 54)",
                                                       " (in 'sem_resid_rs', line 273, column 8 to column 54)",
                                                       " (in 'sem_resid_rs', line 276, column 12 to column 26)",
@@ -307,7 +307,7 @@ static const std::vector<string> locations_array__ = {" (found before start of p
                                                       " (in 'sem_resid_rs', line 141, column 4 to column 29)",
                                                       " (in 'sem_resid_rs', line 142, column 11 to column 13)",
                                                       " (in 'sem_resid_rs', line 142, column 4 to column 25)",
-                                                      " (in 'sem_resid_rs', line 144, column 6 to column 43)",
+                                                      " (in 'sem_resid_rs', line 144, column 6 to column 37)",
                                                       " (in 'sem_resid_rs', line 146, column 40 to column 54)",
                                                       " (in 'sem_resid_rs', line 146, column 8 to column 54)",
                                                       " (in 'sem_resid_rs', line 149, column 12 to column 26)",
@@ -378,19 +378,19 @@ static const std::vector<string> locations_array__ = {" (found before start of p
                                                       " (in 'sem_resid_rs', line 35, column 2 to column 19)",
                                                       " (in 'sem_resid_rs', line 36, column 2 to column 9)",
                                                       " (in 'sem_resid_rs', line 37, column 2 to column 10)",
-                                                      " (in 'sem_resid_rs', line 38, column 8 to column 11)",
-                                                      " (in 'sem_resid_rs', line 38, column 2 to column 30)",
+                                                      " (in 'sem_resid_rs', line 38, column 16 to column 19)",
+                                                      " (in 'sem_resid_rs', line 38, column 2 to column 24)",
                                                       " (in 'sem_resid_rs', line 39, column 9 to column 11)",
                                                       " (in 'sem_resid_rs', line 39, column 13 to column 15)",
                                                       " (in 'sem_resid_rs', line 39, column 2 to column 33)",
-                                                      " (in 'sem_resid_rs', line 40, column 8 to column 10)",
-                                                      " (in 'sem_resid_rs', line 40, column 2 to column 24)",
+                                                      " (in 'sem_resid_rs', line 40, column 14 to column 16)",
+                                                      " (in 'sem_resid_rs', line 40, column 2 to column 18)",
                                                       " (in 'sem_resid_rs', line 41, column 2 to column 14)",
-                                                      " (in 'sem_resid_rs', line 42, column 8 to column 15)",
-                                                      " (in 'sem_resid_rs', line 42, column 2 to column 35)",
-                                                      " (in 'sem_resid_rs', line 43, column 8 to column 10)",
-                                                      " (in 'sem_resid_rs', line 43, column 12 to column 14)",
-                                                      " (in 'sem_resid_rs', line 43, column 2 to column 33)",
+                                                      " (in 'sem_resid_rs', line 42, column 17 to column 24)",
+                                                      " (in 'sem_resid_rs', line 42, column 2 to column 29)",
+                                                      " (in 'sem_resid_rs', line 43, column 19 to column 21)",
[…]
