+++ src/stanExports_cfa_resid_rs.h
@@ -331,13 +331,13 @@ static const std::vector<string> locations_array__ = {" (found before start of p
                                                       " (in 'cfa_resid_rs', line 35, column 2 to column 19)",
                                                       " (in 'cfa_resid_rs', line 36, column 2 to column 20)",
                                                       " (in 'cfa_resid_rs', line 37, column 2 to column 21)",
-                                                      " (in 'cfa_resid_rs', line 38, column 8 to column 11)",
-                                                      " (in 'cfa_resid_rs', line 38, column 2 to column 30)",
+                                                      " (in 'cfa_resid_rs', line 38, column 16 to column 19)",
+                                                      " (in 'cfa_resid_rs', line 38, column 2 to column 24)",
                                                       " (in 'cfa_resid_rs', line 39, column 9 to column 11)",
                                                       " (in 'cfa_resid_rs', line 39, column 13 to column 15)",
                                                       " (in 'cfa_resid_rs', line 39, column 2 to column 33)",
-                                                      " (in 'cfa_resid_rs', line 40, column 8 to column 10)",
-                                                      " (in 'cfa_resid_rs', line 40, column 2 to column 24)",
+                                                      " (in 'cfa_resid_rs', line 40, column 14 to column 16)",
+                                                      " (in 'cfa_resid_rs', line 40, column 2 to column 18)",
                                                       " (in 'cfa_resid_rs', line 41, column 2 to column 37)",
                                                       " (in 'cfa_resid_rs', line 42, column 2 to column 30)",
                                                       " (in 'cfa_resid_rs', line 43, column 2 to column 25)",
