+++ src/stanExports_meta_cfa_resid_rs.h
@@ -268,8 +268,8 @@ static const std::vector<string> locations_array__ = {" (found before start of p
                                                       " (in 'meta_cfa_resid_rs', line 180, column 23 to line 189, column 7)",
                                                       " (in 'meta_cfa_resid_rs', line 180, column 6 to line 189, column 7)",
                                                       " (in 'meta_cfa_resid_rs', line 144, column 4 to line 190, column 5)",
-                                                      " (in 'meta_cfa_resid_rs', line 192, column 12 to column 31)",
-                                                      " (in 'meta_cfa_resid_rs', line 192, column 6 to column 42)",
+                                                      " (in 'meta_cfa_resid_rs', line 192, column 15 to column 34)",
+                                                      " (in 'meta_cfa_resid_rs', line 192, column 6 to column 36)",
                                                       " (in 'meta_cfa_resid_rs', line 193, column 6 to column 17)",
                                                       " (in 'meta_cfa_resid_rs', line 194, column 6 to column 20)",
                                                       " (in 'meta_cfa_resid_rs', line 197, column 10 to column 25)",
@@ -309,22 +309,22 @@ static const std::vector<string> locations_array__ = {" (found before start of p
                                                       " (in 'meta_cfa_resid_rs', line 191, column 4 to line 230, column 5)",
                                                       " (in 'meta_cfa_resid_rs', line 137, column 2 to line 231, column 3)",
                                                       " (in 'meta_cfa_resid_rs', line 36, column 2 to column 20)",
-                                                      " (in 'meta_cfa_resid_rs', line 37, column 8 to column 10)",
-                                                      " (in 'meta_cfa_resid_rs', line 37, column 2 to column 30)",
+                                                      " (in 'meta_cfa_resid_rs', line 37, column 20 to column 22)",
+                                                      " (in 'meta_cfa_resid_rs', line 37, column 2 to column 24)",
                                                       " (in 'meta_cfa_resid_rs', line 38, column 2 to column 20)",
-                                                      " (in 'meta_cfa_resid_rs', line 39, column 8 to column 10)",
                                                       " (in 'meta_cfa_resid_rs', line 39, column 19 to column 21)",
-                                                      " (in 'meta_cfa_resid_rs', line 39, column 23 to column 25)",
-                                                      " (in 'meta_cfa_resid_rs', line 39, column 2 to column 29)",
+                                                      " (in 'meta_cfa_resid_rs', line 39, column 9 to column 11)",
+                                                      " (in 'meta_cfa_resid_rs', line 39, column 13 to column 15)",
+                                                      " (in 'meta_cfa_resid_rs', line 39, column 2 to column 23)",
                                                       " (in 'meta_cfa_resid_rs', line 40, column 2 to column 20)",
                                                       " (in 'meta_cfa_resid_rs', line 41, column 2 to column 21)",
-                                                      " (in 'meta_cfa_resid_rs', line 42, column 8 to column 11)",
-                                                      " (in 'meta_cfa_resid_rs', line 42, column 2 to column 30)",
+                                                      " (in 'meta_cfa_resid_rs', line 42, column 16 to column 19)",
+                                                      " (in 'meta_cfa_resid_rs', line 42, column 2 to column 24)",
                                                       " (in 'meta_cfa_resid_rs', line 43, column 9 to column 11)",
                                                       " (in 'meta_cfa_resid_rs', line 43, column 13 to column 15)",
                                                       " (in 'meta_cfa_resid_rs', line 43, column 2 to column 33)",
-                                                      " (in 'meta_cfa_resid_rs', line 44, column 8 to column 10)",
-                                                      " (in 'meta_cfa_resid_rs', line 44, column 2 to column 24)",
[…]
