+++ src/stanExports_cfa_resid_rs.h
@@ -87,140 +87,140 @@ static const std::vector<string> locations_array__ = {" (found before start of p
                                                       " (in 'cfa_resid_rs', line 91, column 2 to column 58)",
                                                       " (in 'cfa_resid_rs', line 92, column 2 to column 58)",
                                                       " (in 'cfa_resid_rs', line 93, column 2 to column 28)",
-                                                      " (in 'cfa_resid_rs', line 189, column 2 to column 13)",
                                                       " (in 'cfa_resid_rs', line 190, column 2 to column 13)",
-                                                      " (in 'cfa_resid_rs', line 191, column 2 to column 33)",
-                                                      " (in 'cfa_resid_rs', line 192, column 2 to column 26)",
-                                                      " (in 'cfa_resid_rs', line 193, column 2 to column 50)",
-                                                      " (in 'cfa_resid_rs', line 194, column 2 to column 85)",
-                                                      " (in 'cfa_resid_rs', line 195, column 2 to column 39)",
-                                                      " (in 'cfa_resid_rs', line 196, column 2 to column 43)",
-                                                      " (in 'cfa_resid_rs', line 197, column 2 to column 22)",
-                                                      " (in 'cfa_resid_rs', line 198, column 2 to column 49)",
-                                                      " (in 'cfa_resid_rs', line 202, column 4 to column 18)",
-                                                      " (in 'cfa_resid_rs', line 201, column 9 to line 203, column 3)",
-                                                      " (in 'cfa_resid_rs', line 200, column 4 to column 27)",
-                                                      " (in 'cfa_resid_rs', line 199, column 21 to line 201, column 3)",
-                                                      " (in 'cfa_resid_rs', line 199, column 2 to line 203, column 3)",
-                                                      " (in 'cfa_resid_rs', line 209, column 4 to line 211, column 6)",
-                                                      " (in 'cfa_resid_rs', line 208, column 26 to line 212, column 3)",
-                                                      " (in 'cfa_resid_rs', line 208, column 9 to line 212, column 3)",
-                                                      " (in 'cfa_resid_rs', line 207, column 4 to column 29)",
-                                                      " (in 'cfa_resid_rs', line 206, column 26 to line 208, column 3)",
-                                                      " (in 'cfa_resid_rs', line 206, column 9 to line 212, column 3)",
-                                                      " (in 'cfa_resid_rs', line 205, column 4 to column 24)",
-                                                      " (in 'cfa_resid_rs', line 204, column 19 to line 206, column 3)",
-                                                      " (in 'cfa_resid_rs', line 204, column 2 to line 212, column 3)",
-                                                      " (in 'cfa_resid_rs', line 214, column 4 to column 16)",
-                                                      " (in 'cfa_resid_rs', line 217, column 8 to column 17)",
-                                                      " (in 'cfa_resid_rs', line 218, column 8 to column 49)",
-                                                      " (in 'cfa_resid_rs', line 219, column 8 to column 34)",
-                                                      " (in 'cfa_resid_rs', line 216, column 27 to line 220, column 7)",
-                                                      " (in 'cfa_resid_rs', line 216, column 6 to line 220, column 7)",
-                                                      " (in 'cfa_resid_rs', line 215, column 20 to line 221, column 5)",
-                                                      " (in 'cfa_resid_rs', line 215, column 4 to line 221, column 5)",
-                                                      " (in 'cfa_resid_rs', line 213, column 19 to line 222, column 3)",
-                                                      " (in 'cfa_resid_rs', line 213, column 2 to line 222, column 3)",
-                                                      " (in 'cfa_resid_rs', line 224, column 4 to column 11)",
-                                                      " (in 'cfa_resid_rs', line 225, column 11 to column 13)",
[…]
