+++ src/stanExports_sem_resid_rs.h
@@ -87,165 +87,165 @@ static const std::vector<string> locations_array__ = {" (found before start of p
                                                       " (in 'sem_resid_rs', line 96, column 2 to column 27)",
                                                       " (in 'sem_resid_rs', line 97, column 2 to column 47)",
                                                       " (in 'sem_resid_rs', line 98, column 2 to column 28)",
-                                                      " (in 'sem_resid_rs', line 216, column 2 to column 13)",
                                                       " (in 'sem_resid_rs', line 217, column 2 to column 13)",
-                                                      " (in 'sem_resid_rs', line 218, column 2 to column 33)",
-                                                      " (in 'sem_resid_rs', line 219, column 2 to column 26)",
-                                                      " (in 'sem_resid_rs', line 220, column 2 to column 52)",
+                                                      " (in 'sem_resid_rs', line 218, column 2 to column 13)",
+                                                      " (in 'sem_resid_rs', line 219, column 2 to column 33)",
+                                                      " (in 'sem_resid_rs', line 220, column 2 to column 26)",
                                                       " (in 'sem_resid_rs', line 221, column 2 to column 52)",
-                                                      " (in 'sem_resid_rs', line 222, column 2 to column 50)",
+                                                      " (in 'sem_resid_rs', line 222, column 2 to column 52)",
                                                       " (in 'sem_resid_rs', line 223, column 2 to column 50)",
-                                                      " (in 'sem_resid_rs', line 224, column 2 to column 38)",
-                                                      " (in 'sem_resid_rs', line 225, column 2 to column 22)",
-                                                      " (in 'sem_resid_rs', line 226, column 2 to column 47)",
-                                                      " (in 'sem_resid_rs', line 227, column 2 to column 39)",
-                                                      " (in 'sem_resid_rs', line 228, column 2 to column 43)",
-                                                      " (in 'sem_resid_rs', line 229, column 2 to column 22)",
-                                                      " (in 'sem_resid_rs', line 230, column 2 to column 49)",
-                                                      " (in 'sem_resid_rs', line 234, column 4 to column 18)",
-                                                      " (in 'sem_resid_rs', line 233, column 9 to line 235, column 3)",
-                                                      " (in 'sem_resid_rs', line 232, column 4 to column 27)",
-                                                      " (in 'sem_resid_rs', line 231, column 21 to line 233, column 3)",
-                                                      " (in 'sem_resid_rs', line 231, column 2 to line 235, column 3)",
-                                                      " (in 'sem_resid_rs', line 241, column 4 to line 243, column 6)",
-                                                      " (in 'sem_resid_rs', line 240, column 26 to line 244, column 3)",
-                                                      " (in 'sem_resid_rs', line 240, column 9 to line 244, column 3)",
-                                                      " (in 'sem_resid_rs', line 239, column 4 to column 29)",
-                                                      " (in 'sem_resid_rs', line 238, column 26 to line 240, column 3)",
-                                                      " (in 'sem_resid_rs', line 238, column 9 to line 244, column 3)",
-                                                      " (in 'sem_resid_rs', line 237, column 4 to column 24)",
-                                                      " (in 'sem_resid_rs', line 236, column 19 to line 238, column 3)",
-                                                      " (in 'sem_resid_rs', line 236, column 2 to line 244, column 3)",
-                                                      " (in 'sem_resid_rs', line 246, column 4 to column 16)",
-                                                      " (in 'sem_resid_rs', line 249, column 8 to column 17)",
-                                                      " (in 'sem_resid_rs', line 250, column 8 to column 49)",
-                                                      " (in 'sem_resid_rs', line 251, column 8 to column 34)",
[…]
