+++ src/stanExports_meta_cfa_resid_rs.h
@@ -118,60 +118,63 @@ static const std::vector<string> locations_array__ = {" (found before start of p
                                                       " (in 'meta_cfa_resid_rs', line 251, column 2 to line 259, column 3)",
                                                       " (in 'meta_cfa_resid_rs', line 261, column 11 to column 13)",
                                                       " (in 'meta_cfa_resid_rs', line 261, column 4 to column 54)",
-                                                      " (in 'meta_cfa_resid_rs', line 262, column 4 to column 52)",
-                                                      " (in 'meta_cfa_resid_rs', line 263, column 4 to column 28)",
-                                                      " (in 'meta_cfa_resid_rs', line 264, column 4 to column 26)",
-                                                      " (in 'meta_cfa_resid_rs', line 266, column 6 to column 83)",
-                                                      " (in 'meta_cfa_resid_rs', line 265, column 19 to line 267, column 5)",
-                                                      " (in 'meta_cfa_resid_rs', line 265, column 4 to line 267, column 5)",
-                                                      " (in 'meta_cfa_resid_rs', line 260, column 17 to line 268, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 260, column 2 to line 268, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 270, column 4 to column 16)",
-                                                      " (in 'meta_cfa_resid_rs', line 273, column 8 to column 17)",
-                                                      " (in 'meta_cfa_resid_rs', line 274, column 8 to column 49)",
-                                                      " (in 'meta_cfa_resid_rs', line 275, column 8 to column 34)",
-                                                      " (in 'meta_cfa_resid_rs', line 272, column 27 to line 276, column 7)",
-                                                      " (in 'meta_cfa_resid_rs', line 272, column 6 to line 276, column 7)",
-                                                      " (in 'meta_cfa_resid_rs', line 271, column 20 to line 277, column 5)",
-                                                      " (in 'meta_cfa_resid_rs', line 271, column 4 to line 277, column 5)",
-                                                      " (in 'meta_cfa_resid_rs', line 269, column 19 to line 278, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 269, column 2 to line 278, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 280, column 4 to column 16)",
-                                                      " (in 'meta_cfa_resid_rs', line 281, column 4 to column 24)",
-                                                      " (in 'meta_cfa_resid_rs', line 288, column 10 to column 27)",
-                                                      " (in 'meta_cfa_resid_rs', line 289, column 10 to column 85)",
-                                                      " (in 'meta_cfa_resid_rs', line 287, column 39 to line 290, column 9)",
-                                                      " (in 'meta_cfa_resid_rs', line 287, column 15 to line 290, column 9)",
-                                                      " (in 'meta_cfa_resid_rs', line 285, column 10 to column 19)",
-                                                      " (in 'meta_cfa_resid_rs', line 286, column 10 to column 41)",
-                                                      " (in 'meta_cfa_resid_rs', line 284, column 40 to line 287, column 9)",
-                                                      " (in 'meta_cfa_resid_rs', line 284, column 8 to line 290, column 9)",
-                                                      " (in 'meta_cfa_resid_rs', line 283, column 22 to line 291, column 7)",
[…]
