+++ NAMESPACE
@@ -1,12 +1,12 @@
 # Generated by roxygen2: do not edit by hand
 
+export(casewise_log_likelihood)
 export(init_minorbsem)
 export(meta_mbcfa)
 export(minorbsem)
 export(new_mbsempriors)
 export(parameter_hist)
 export(parameter_trace)
-export(plot)
 export(plot_residuals)
 export(pretty_print_summary)
 exportClasses(mbsempriors)
