+++ src/RcppExports.cpp
@@ -1,6 +1,7 @@
 // Generated by using Rcpp::compileAttributes() -> do not edit by hand
 // Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+#include <RcppArmadillo.h>
 #include <RcppEigen.h>
 #include <Rcpp.h>
 
@@ -11,12 +12,42 @@ Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
 Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
+// ldmvnrm_arma
+arma::rowvec ldmvnrm_arma(arma::mat const& x, arma::colvec const& mean, arma::mat const& sigma, int const& n, int const& xdim);
+RcppExport SEXP _minorbsem_ldmvnrm_arma(SEXP xSEXP, SEXP meanSEXP, SEXP sigmaSEXP, SEXP nSEXP, SEXP xdimSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat const& >::type x(xSEXP);
+    Rcpp::traits::input_parameter< arma::colvec const& >::type mean(meanSEXP);
+    Rcpp::traits::input_parameter< arma::mat const& >::type sigma(sigmaSEXP);
+    Rcpp::traits::input_parameter< int const& >::type n(nSEXP);
+    Rcpp::traits::input_parameter< int const& >::type xdim(xdimSEXP);
+    rcpp_result_gen = Rcpp::wrap(ldmvnrm_arma(x, mean, sigma, n, xdim));
+    return rcpp_result_gen;
+END_RCPP
+}
+// ldmvnrm_list_arma
+arma::mat ldmvnrm_list_arma(arma::mat const& x, arma::colvec const& mean, arma::mat const& sigma_list);
+RcppExport SEXP _minorbsem_ldmvnrm_list_arma(SEXP xSEXP, SEXP meanSEXP, SEXP sigma_listSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat const& >::type x(xSEXP);
+    Rcpp::traits::input_parameter< arma::colvec const& >::type mean(meanSEXP);
+    Rcpp::traits::input_parameter< arma::mat const& >::type sigma_list(sigma_listSEXP);
+    rcpp_result_gen = Rcpp::wrap(ldmvnrm_list_arma(x, mean, sigma_list));
+    return rcpp_result_gen;
+END_RCPP
+}
 
 RcppExport SEXP _rcpp_module_boot_stan_fit4cfa_resid_rs_mod();
 RcppExport SEXP _rcpp_module_boot_stan_fit4meta_cfa_resid_rs_mod();
 RcppExport SEXP _rcpp_module_boot_stan_fit4sem_resid_rs_mod();
 
 static const R_CallMethodDef CallEntries[] = {
+    {"_minorbsem_ldmvnrm_arma", (DL_FUNC) &_minorbsem_ldmvnrm_arma, 5},
+    {"_minorbsem_ldmvnrm_list_arma", (DL_FUNC) &_minorbsem_ldmvnrm_list_arma, 3},
[…]
