+++ R/meta_mbcfa.R
@@ -81,7 +81,8 @@ meta_mbcfa <- function(
     ncores = max(parallel::detectCores() - 2, 1),
     priors = new_mbsempriors(),
     show = TRUE,
-    show_messages = TRUE) {
+    show_messages = TRUE,
+    target = "rstan") {
   message("Processing user input ...")
 
   # Model cannot be NULL
@@ -100,6 +101,9 @@ meta_mbcfa <- function(
 
   # Must provide either data and group or sample_cov and sample_nobs
   user_input_check("data-meta", data, group, sample_cov, sample_nobs)
+
+  # target must be valid
+  user_input_check("target", target)
 
   # Run lavaan fit
   if (!is.null(data)) {
@@ -127,31 +131,9 @@ meta_mbcfa <- function(
 
   message("User input fully processed :)\n Now to modeling.")
 
-  if (data_list$sem_indicator == 0) {
-    mod_resid <- stanmodels$meta_cfa_resid_rs
-  }
-
-  message("Fitting Stan model ...")
-
-  stan_fit <- rstan::sampling(
-    mod_resid,
-    data = data_list,
-    chains = chains,
-    cores = ncores,
-    seed = seed,
-    warmup = warmup,
-    iter = warmup + sampling,
-    refresh = refresh,
-    init = function() {
-      list(
-        resids = rep(1e-3, (data_list$Ni^2 - data_list$Ni) / 2)
-      )
-    },
-    control = list(
-      adapt_delta = adapt_delta,
-      max_treedepth = max_treedepth
-    ),
-    show_messages = show_messages
+  stan_fit <- target_fitter(
+    target, data_list, seed, warmup, sampling, refresh,
+    adapt_delta, max_treedepth, chains, ncores, show_messages
   )
 
   mbsem_results <- clean_up_stan_fit(
