+++ R/RcppExports.R
@@ -1,0 +1,11 @@
+# Generated by using Rcpp::compileAttributes() -> do not edit by hand
+# Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
+
+ldmvnrm_arma <- function(x, mean, sigma, n, xdim) {
+    .Call(`_minorbsem_ldmvnrm_arma`, x, mean, sigma, n, xdim)
+}
+
+ldmvnrm_list_arma <- function(x, mean, sigma_list) {
+    .Call(`_minorbsem_ldmvnrm_list_arma`, x, mean, sigma_list)
+}
+
