+++ src/stanExports_meta_cfa_resid_rs.h
@@ -98,83 +98,83 @@ static const std::vector<string> locations_array__ = {" (found before start of p
                                                       " (in 'meta_cfa_resid_rs', line 239, column 2 to column 22)",
                                                       " (in 'meta_cfa_resid_rs', line 240, column 2 to column 49)",
                                                       " (in 'meta_cfa_resid_rs', line 241, column 2 to column 47)",
-                                                      " (in 'meta_cfa_resid_rs', line 242, column 2 to column 17)",
-                                                      " (in 'meta_cfa_resid_rs', line 243, column 2 to column 18)",
-                                                      " (in 'meta_cfa_resid_rs', line 244, column 2 to column 29)",
-                                                      " (in 'meta_cfa_resid_rs', line 245, column 2 to column 23)",
-                                                      " (in 'meta_cfa_resid_rs', line 249, column 4 to column 18)",
-                                                      " (in 'meta_cfa_resid_rs', line 248, column 9 to line 250, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 247, column 4 to column 27)",
-                                                      " (in 'meta_cfa_resid_rs', line 246, column 21 to line 248, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 246, column 2 to line 250, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 256, column 4 to line 258, column 6)",
-                                                      " (in 'meta_cfa_resid_rs', line 255, column 26 to line 259, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 255, column 9 to line 259, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 254, column 4 to column 29)",
-                                                      " (in 'meta_cfa_resid_rs', line 253, column 26 to line 255, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 253, column 9 to line 259, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 252, column 4 to column 24)",
-                                                      " (in 'meta_cfa_resid_rs', line 251, column 19 to line 253, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 251, column 2 to line 259, column 3)",
+                                                      " (in 'meta_cfa_resid_rs', line 242, column 2 to column 18)",
+                                                      " (in 'meta_cfa_resid_rs', line 243, column 2 to column 29)",
+                                                      " (in 'meta_cfa_resid_rs', line 244, column 2 to column 23)",
+                                                      " (in 'meta_cfa_resid_rs', line 248, column 4 to column 18)",
+                                                      " (in 'meta_cfa_resid_rs', line 247, column 9 to line 249, column 3)",
+                                                      " (in 'meta_cfa_resid_rs', line 246, column 4 to column 27)",
+                                                      " (in 'meta_cfa_resid_rs', line 245, column 21 to line 247, column 3)",
+                                                      " (in 'meta_cfa_resid_rs', line 245, column 2 to line 249, column 3)",
+                                                      " (in 'meta_cfa_resid_rs', line 255, column 4 to line 257, column 6)",
+                                                      " (in 'meta_cfa_resid_rs', line 254, column 26 to line 258, column 3)",
+                                                      " (in 'meta_cfa_resid_rs', line 254, column 9 to line 258, column 3)",
+                                                      " (in 'meta_cfa_resid_rs', line 253, column 4 to column 29)",
+                                                      " (in 'meta_cfa_resid_rs', line 252, column 26 to line 254, column 3)",
+                                                      " (in 'meta_cfa_resid_rs', line 252, column 9 to line 258, column 3)",
[…]
