+++ src/stanExports_meta_cfa_resid_rs.h
@@ -90,227 +90,230 @@ static const std::vector<string> locations_array__ = {" (found before start of p
                                                       " (in 'meta_cfa_resid_rs', line 102, column 2 to column 22)",
                                                       " (in 'meta_cfa_resid_rs', line 103, column 2 to column 50)",
                                                       " (in 'meta_cfa_resid_rs', line 104, column 2 to column 43)",
-                                                      " (in 'meta_cfa_resid_rs', line 234, column 2 to column 26)",
-                                                      " (in 'meta_cfa_resid_rs', line 235, column 2 to column 50)",
-                                                      " (in 'meta_cfa_resid_rs', line 236, column 2 to column 85)",
-                                                      " (in 'meta_cfa_resid_rs', line 237, column 2 to column 39)",
-                                                      " (in 'meta_cfa_resid_rs', line 238, column 2 to column 43)",
-                                                      " (in 'meta_cfa_resid_rs', line 239, column 2 to column 22)",
-                                                      " (in 'meta_cfa_resid_rs', line 240, column 2 to column 49)",
-                                                      " (in 'meta_cfa_resid_rs', line 241, column 2 to column 47)",
-                                                      " (in 'meta_cfa_resid_rs', line 242, column 2 to column 18)",
-                                                      " (in 'meta_cfa_resid_rs', line 243, column 2 to column 29)",
-                                                      " (in 'meta_cfa_resid_rs', line 244, column 2 to column 23)",
-                                                      " (in 'meta_cfa_resid_rs', line 248, column 4 to column 18)",
-                                                      " (in 'meta_cfa_resid_rs', line 247, column 9 to line 249, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 246, column 4 to column 27)",
-                                                      " (in 'meta_cfa_resid_rs', line 245, column 21 to line 247, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 245, column 2 to line 249, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 255, column 4 to line 257, column 6)",
-                                                      " (in 'meta_cfa_resid_rs', line 254, column 26 to line 258, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 254, column 9 to line 258, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 253, column 4 to column 29)",
-                                                      " (in 'meta_cfa_resid_rs', line 252, column 26 to line 254, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 252, column 9 to line 258, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 251, column 4 to column 24)",
-                                                      " (in 'meta_cfa_resid_rs', line 250, column 19 to line 252, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 250, column 2 to line 258, column 3)",
-                                                      " (in 'meta_cfa_resid_rs', line 260, column 11 to column 13)",
-                                                      " (in 'meta_cfa_resid_rs', line 260, column 4 to column 54)",
-                                                      " (in 'meta_cfa_resid_rs', line 261, column 11 to column 13)",
-                                                      " (in 'meta_cfa_resid_rs', line 261, column 4 to column 34)",
-                                                      " (in 'meta_cfa_resid_rs', line 262, column 4 to column 22)",
-                                                      " (in 'meta_cfa_resid_rs', line 263, column 4 to column 28)",
-                                                      " (in 'meta_cfa_resid_rs', line 264, column 4 to column 26)",
-                                                      " (in 'meta_cfa_resid_rs', line 266, column 6 to column 60)",
[…]
