+++ README.Rmd
@@ -19,6 +19,7 @@ devtools::load_all()
 [![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/minorbsem)](https://cran.r-project.org/package=minorbsem)
 ![GitHub last commit](https://img.shields.io/github/last-commit/jamesuanhoro/minorbsem)
 [![R-CMD-check](https://github.com/jamesuanhoro/minorbsem/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/jamesuanhoro/minorbsem/actions/workflows/R-CMD-check.yaml)
+[![minorbsem status badge](https://jamesuanhoro.r-universe.dev/badges/minorbsem)](https://jamesuanhoro.r-universe.dev)
 <!-- badges: end -->
 
 #### Table of Contents
@@ -27,8 +28,6 @@ devtools::load_all()
   - [Goals](#goals)
   - [Permitted models and supported data types](#permitted-models-and-supported-data-types)
 - [Installation](#installation)
-  - [Install remotes](#install-remotes)
-  - [Install minorbsem](#install-minorbsem)
 - [A reasonably complete demonstration](#a-reasonably-complete-demonstration)
   - [Model comparisons](#model-comparisons)
 - [Additional examples](#additional-examples)
@@ -74,25 +73,15 @@ All data are assumed multivariate normal, i.e. no binary, ordinal models.
 
 ## Installation
 
-### Install remotes
-
-Restart R prior to installation. minorbsem requires that R is at least version
-3.4.0. `version` command shows the installed R version.
-If the installed version is less than 3.4.0, then update R before proceeding.
-
-Install `remotes` to allow installation of R packages that are hosted on GitHub
-instead of Cran (the default repository for R packages):
-
-``` r
-install.packages("remotes")
-```
-
-### Install minorbsem
-
 Install minorbsem:
 
 ``` r
-remotes::install_github("jamesuanhoro/minorbsem")
+install.packages(
+  'minorbsem',
+  repos = c(
+    'https://jamesuanhoro.r-universe.dev', 'https://cloud.r-project.org'
+  )
+)
 ```
 
 ## A reasonably complete demonstration
