+++ src/stanExports_sem_resid_rs.h
@@ -1,3400 +1,0 @@
-// Generated by rstantools.  Do not edit by hand.
-
-/*
-    minorbsem is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    minorbsem is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with minorbsem.  If not, see <http://www.gnu.org/licenses/>.
-*/
-#ifndef MODELS_HPP
-#define MODELS_HPP
-#define STAN__SERVICES__COMMAND_HPP
-#ifndef USE_STANC3
-#define USE_STANC3
-#endif
-#include <rstan/rstaninc.hpp>
-// Code generated by %%NAME%% %%VERSION%%
-#include <stan/model/model_header.hpp>
-namespace model_sem_resid_rs_namespace {
-inline void validate_positive_index(const char* var_name, const char* expr,
-                                    int val) {
-  if (val < 1) {
-    std::stringstream msg;
-    msg << "Found dimension size less than one in simplex declaration"
-        << "; variable=" << var_name << "; dimension size expression=" << expr
-        << "; expression value=" << val;
-    std::string msg_str(msg.str());
-    throw std::invalid_argument(msg_str.c_str());
-  }
-}
-inline void validate_unit_vector_index(const char* var_name, const char* expr,
-                                       int val) {
-  if (val <= 1) {
-    std::stringstream msg;
-    if (val == 1) {
-      msg << "Found dimension size one in unit vector declaration."
-          << " One-dimensional unit vector is discrete"
-          << " but the target distribution must be continuous."
-          << " variable=" << var_name << "; dimension size expression=" << expr;
-    } else {
-      msg << "Found dimension size less than one in unit vector declaration"
-          << "; variable=" << var_name << "; dimension size expression=" << expr
-          << "; expression value=" << val;
-    }
-    std::string msg_str(msg.str());
-    throw std::invalid_argument(msg_str.c_str());
-  }
-}
-using std::istream;
-using std::string;
-using std::stringstream;
-using std::vector;
-using std::pow;
-using stan::io::dump;
-using stan::math::lgamma;
-using stan::model::model_base_crtp;
-using stan::model::rvalue;
-using stan::model::cons_list;
-using stan::model::index_uni;
-using stan::model::index_max;
-using stan::model::index_min;
-using stan::model::index_min_max;
-using stan::model::index_multi;
-using stan::model::index_omni;
-using stan::model::nil_index_list;
-using namespace stan::math;
-using stan::math::pow; 
-stan::math::profile_map profiles__;
-static int current_statement__= 0;
-static const std::vector<string> locations_array__ = {" (found before start of program)",
-                                                      " (in 'sem_resid_rs', line 87, column 2 to column 39)",
-                                                      " (in 'sem_resid_rs', line 88, column 2 to column 41)",
-                                                      " (in 'sem_resid_rs', line 89, column 2 to column 24)",
-                                                      " (in 'sem_resid_rs', line 90, column 2 to column 38)",
-                                                      " (in 'sem_resid_rs', line 91, column 2 to column 33)",
-                                                      " (in 'sem_resid_rs', line 92, column 2 to column 31)",
-                                                      " (in 'sem_resid_rs', line 93, column 2 to column 32)",
-                                                      " (in 'sem_resid_rs', line 94, column 2 to column 51)",
-                                                      " (in 'sem_resid_rs', line 95, column 2 to column 47)",
[…]
