### 2023-04-11 b01ce63
A number of changes
- Switched from kableExtra to huxtable
- Modified tests accordingly
- Replaced all regexpr search with grepl
- Showing additional RMSEAs in type = "dep" plots
