+++ tests/testthat/setup.R
@@ -45,24 +45,23 @@ mbsem_test_plot_residuals <- function(fit, method) {
   }
 }
 
-mbsem_test_kbls_shared <- function(kbl, method, meta = FALSE) {
-  testthat::expect_true(inherits(kbl, "kableExtra"))
-  testthat::expect_true(regexpr(method, kbl, ignore.case = TRUE) > 0)
-  testthat::expect_true(regexpr("Goodness of fit", kbl, ignore.case = TRUE) > 0)
-  testthat::expect_true(regexpr("RMSE", kbl, ignore.case = TRUE) > 0)
-  testthat::expect_true(regexpr("Factor loadings", kbl, ignore.case = TRUE) > 0)
+mbsem_test_pp_shared <- function(print_out, method, meta = FALSE) {
+  testthat::expect_true(grepl(method, print_out, ignore.case = TRUE))
+  testthat::expect_true(grepl("Goodness of fit", print_out, ignore.case = TRUE))
+  testthat::expect_true(grepl("RMSE", print_out, ignore.case = TRUE))
+  testthat::expect_true(grepl("Factor loadings", print_out, ignore.case = TRUE))
   if (isTRUE(meta)) {
-    testthat::expect_true(regexpr("RMSEA", kbl, ignore.case = TRUE) > 0)
+    testthat::expect_true(grepl("RMSEA", print_out, ignore.case = TRUE))
   } else {
-    testthat::expect_true(regexpr("PPP", kbl, ignore.case = TRUE) > 0)
+    testthat::expect_true(grepl("PPP", print_out, ignore.case = TRUE))
   }
 }
 
 dat_cov <- function(dat = "HS") {
   if (dat == "HS") {
-    dat <- HS[, paste0("x", 1:9)]
+    dat <- HS[, paste0("x", 1:9)] # nolint
   } else if (dat == "PD") {
-    dat <- PD
+    dat <- PD # nolint
   }
 
   input <- list(data = NULL, cov = NULL, nobs = NULL)
