### 2023-04-11 937509c
Updated documentation and fixed minor error
