+++ inst/REFERENCES.bib
@@ -1,11 +1,19 @@
 
 @article{uanhoro_modeling_2023,
-  title    = {Modeling misspecification as a parameter in {Bayesian} structural equation modeling},
-  abstract = {Accounting for model misspecification in Bayesian structural equation models is an active area of research. We present a uniquely Bayesian approach to misspecification that models the degree of misspecification as a parameter – a parameter akin to the correlation root mean squared residual. The misspecification parameter can be interpreted on its own terms as a measure of absolute model fit, and allows for comparing different models fit to the same data. By estimating the degree of misspecification simultaneously with structural parameters, the uncertainty about structural parameters reflect the degree of model misspecification. This results in a model that produces more reliable inference than extant Bayesian SEMs. Additionally, the approach estimates the residual covariance matrix which can be the basis for diagnosing misspecifications and updating a hypothesized model. These features are confirmed using simulation studies. Demonstrations with a variety of real world examples show additional properties of the approach.},
+  title    = {Modeling Misspecification as a Parameter in {Bayesian} Structural Equation Models},
+  volume   = 0,
+  number   = {0},
+  issn     = {0013-1644},
+  url      = {https://journals.sagepub.com/doi/abs/10.1177/00131644231165306},
+  doi      = {10.1177/00131644231165306},
   language = {en},
+  urldate  = {2023-04-24},
   journal  = {Educational and Psychological Measurement},
   author   = {Uanhoro, James Ohisei},
-  year     = {forthcoming}
+  month    = apr,
+  year     = {2023},
+  pages    = {00131644231165306},
+  file     = {uanhoro_2023_modeling_misspecification_as_a_parameter_in_bayesian_structural_equation.pdf:/home/james/Dropbox/zotero-library/AUTH/Uanhoro/uanhoro_2023_modeling_misspecification_as_a_parameter_in_bayesian_structural_equation.pdf:application/pdf}
 }
 
 @article{uanhoro_hierarchical_2022,
