### 2023-05-05 c6085c6
Update README.Rmd
minor copy editing
