### 2023-05-05 6ca08de
Update README.Rmd
added some basic information about loo
