### 2023-05-05 #5292 @smasongarrison
# Title: Review of the minorbsem R package paper

I have reviewed the submission: "minorbsem: An R package for structural equation models that account for the influence of minor factors" and found that everything is in order. I was able to run everything and the package performed as expected. I recommend acceptance of this paper. During my review, I issued two minor pull-requests. My acceptance does not hinge on those changes, but I believe they would improve the user experience.

# Minor Notes:

- I would love to see this package made available on CRAN.

## Paper Summary

In the paper summary and features, I suggest explicitly mentioning that this method is for fitting Bayesian SEM (as opposed to frequentist SEM).

## Documentation Summary

I performed some minor copy-editing that I think will make the paper more accessible for the audience, including spelling out a few acronyms. I have issued a pull request (https://github.com/jamesuanhoro/minorbsem/pull/92), as I found it easier to make the changes directly instead of describing them. However, I defer to the author, @jamesuanhoro.

## References

The references are out-of-date on the JOSS branch. They are current on the main branch (https://github.com/jamesuanhoro/minorbsem/commit/6fe387a76fd4c2be91d539874118ec91e6f1e548). To be fair, the paper is currently in press, so I do not fault the author for this. I have issued a pull request (https://github.com/jamesuanhoro/minorbsem/pull/91) to address this issue.
