+++ inst/REFERENCES.bib
@@ -136,12 +136,12 @@
   url       = {https://ggplot2.tidyverse.org}
 }
 
-@manual{kable_ex,
-  title  = {kableExtra: Construct Complex Table with 'kable' and Pipe Syntax},
-  author = {Hao Zhu},
-  year   = {2021},
-  note   = {R package version 1.3.4},
-  url    = {https://CRAN.R-project.org/package=kableExtra}
+@manual{huxtable,
+  title  = {huxtable: Easily Create and Style Tables for LaTeX, HTML and Other Formats},
+  author = {David Hugh-Jones},
+  year   = {2022},
+  note   = {R package version 5.5.2},
+  url    = {https://CRAN.R-project.org/package=huxtable}
 }
 
 @misc{posterior,
