### 2023-05-06 f37a99e
This should Close #8
- fs_par added for factor sd scale parameter.
