+++ DESCRIPTION
@@ -5,8 +5,8 @@ Authors@R: c(
     person(given = c("James", "O."), family = "Uanhoro",
            email = "james.uanhoro@unt.edu", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0002-4843-927X")))
-Description: Structural equation models (SEMs) rarely reject the null hypothesis that
-    there is no model misspecification. One explanation for this problem
+Description: Structural equation models (SEMs) rarely reject the null hypothesis
+    that there is no model misspecification. One explanation for this problem
     is that covariance structures are influenced by major factors which we can hypothesize about
     and minor factors which we cannot predict a-priori. This package lets you fit regular and
     meta-analytic Bayesian SEMs that estimate the influence of minor factors.
