+++ R/01_class_methods.R
@@ -37,7 +37,7 @@ methods::setMethod(
     if (object@data_list$method >= 90) {
       warning(paste0(
         "There are no residuals to plot when ",
-        "method == \"none\", \"WB\", \"WB-cond\"."
+        "method == \"none\", \"WB\", \"WB-cond\", \"WW\"."
       ))
     }
 
