+++ R/casewise_log_likelihood.R
@@ -88,7 +88,7 @@ casewise_log_likelihood <- function(
   if (data_list$method >= 90 && isTRUE(include_residuals)) {
     warn_msg <- paste0(
       "include_residuals = TRUE is ignored when ",
-      "minorbsem method == \"none\", \"WB\", \"WB-cond\"."
+      "minorbsem method == \"none\", \"WB\", \"WB-cond\", \"WW\"."
     )
     warning(warn_msg)
   }
