### 2023-06-13 jamesuanhoro/minorbsem#98 @aaronpeikert
make startup message more actionable
This is currently the result of:
```
> library(minorbsem)

###############################################################################
This is minorbsem 0.2.3
All users of R (or SEM) are invited to submit functions or ideas for functions.
###############################################################################
```

But it is unclear how users should submit such ideas (bugs?). Maybe suggest opening an issue on GH.
