### 2023-06-13 #5292 @aaronpeikert
Sorry for the delay. Finally, I made it through and found almost no issues. Great work @jamesuanhoro!
The only thing that could be a tad clearer is:

E.g. with a line the the readme and startup message https://github.com/jamesuanhoro/minorbsem/issues/98
