### 2023-06-13 43c375c
Contribution guidelines
- Resolves #98
