### 2023-06-13 jamesuanhoro/minorbsem#98 @jamesuanhoro
Thank you Aaron, I put together messages I think work.
