### 2023-06-19 8f1099f
Bayesian SEMs mentioned in summary + updated a paper citation
