+++ paper.md
@@ -20,7 +20,7 @@ bibliography: inst/REFERENCES.bib
 
 # Summary
 
-minorbsem is an R package that allows users to fit structural equation models (SEMs) under the assumption that minor factors influence the relationship between observed indicators. SEMs are models that capture the inter-relations between factors (latent variables) and the relations between these factors and observed items. However, SEMs are rarely able to reject the null hypothesis that there is no misspecification in the hypothesized model. One explanation for this problem is that covariance structures are influenced by major factors that we can hypothesize about, and minor factors that we cannot predict a-priori [@maccallum_representing_1991].
+minorbsem is an R package that allows users to fit Bayesian structural equation models (SEMs) under the assumption that minor factors influence the relationship between observed indicators. SEMs are models that capture the inter-relations between factors (latent variables) and the relations between these factors and observed items. However, SEMs are rarely able to reject the null hypothesis that there is no misspecification in the hypothesized model. One explanation for this problem is that covariance structures are influenced by major factors that we can hypothesize about, and minor factors that we cannot predict a-priori [@maccallum_representing_1991].
 
 minorbsem accounts for the influence of minor factors by estimating all residual covariances between the observed items [@uanhoro_modeling_2023]. These residual covariances are estimated with priors that shrink the residual covariances towards zero. Additionally, the model returns the magnitude of the influence of minor factors. By estimating the influence of minor factors simultaneously with major factors, the parameters of interest (related to major factors) in the SEM will reflect the uncertainty due to the influence of minor factors.
 
