+++ inst/REFERENCES.bib
@@ -17,17 +17,19 @@
 
 @article{uanhoro_hierarchical_2022,
   title    = {Hierarchical Covariance Estimation Approach to Meta-Analytic Structural Equation Modeling},
-  volume   = 0,
+  volume   = {30},
   issn     = {1070-5511},
+  url      = {https://doi.org/10.1080/10705511.2022.2142128},
   doi      = {10.1080/10705511.2022.2142128},
   abstract = {We present a fully Bayesian approach to meta-analytic SEM based on hierarchical modeling of sample covariance matrices. The approach allows for flexible models that would not be identified under a traditional maximum likelihood approach. The approach allows for the inclusion of moderators, produces a global fit index, and permits the investigation of local misspecification. Simulation-based calibration studies show that the Bayesian computation procedure produces valid inferences for commonplace meta-analytic SEM applications. We demonstrate the approach with diverse data analysis examples and provide accompanying R code to support adoption and additional study of the approach. Finally, we lay out proposals that have the potential to extend the approach to accommodate the wide variety of analyses and data conditions that comprise meta-analytic SEM applications.},
-  number   = {0},
+  number   = {4},
+  urldate  = {2023-06-19},
   journal  = {Structural Equation Modeling: A Multidisciplinary Journal},
-  author   = {Uanhoro, James Ohisei},
-  month    = dec,
-  year     = {2022},
+  author   = {Uanhoro, James O.},
+  month    = jul,
+  year     = {2023},
   keywords = {Bayesian estimation, hierarchical covariance, meta-analytic SEM},
-  pages    = {1--15}
+  pages    = {532--546}
 }
 
 @article{maccallum_representing_1991,
