+++ inst/REFERENCES.bib
@@ -25,7 +25,7 @@
   number   = {4},
   urldate  = {2023-06-19},
   journal  = {Structural Equation Modeling: A Multidisciplinary Journal},
-  author   = {Uanhoro, James O.},
+  author   = {Uanhoro, James Ohisei},
   month    = jul,
   year     = {2023},
   keywords = {Bayesian estimation, hierarchical covariance, meta-analytic SEM},
