### 2023-06-19 #5292 @jamesuanhoro
minorbsem version 0.2.4: https://github.com/jamesuanhoro/minorbsem/releases/tag/v0.2.4
Zenodo DOI: https://doi.org/10.5281/zenodo.8057759
