### 2023-06-20 #5292 @samhforbes
Hi @jamesuanhoro do you know why the citations to your work aren't rendering properly in text (giving your full name)? I can't immediately see what's gone wrong here.
