### 2023-06-20 #5292 @samhforbes
@jamesuanhoro I think if I've understood this correctly, the issue is that the ISBN for Muthen 2012 contains the ISSN. Perhaps that field needs to be removed?
