### 2023-06-21 #5292 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
